Rename sequences2.R etc. to sequences.R
